email-service/requirements.txt
functions-framework==3.*
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
google-cloud-storage==2.17.0